backend/pyproject.toml
[tool.poetry]
package-mode = false
description = ""
authors = ["Desunovu <desunovu@gmail.com>"]
license = "CC BY-NC-SA 4.0"
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
SQLAlchemy = "^2.0.30"
pytest = "^8.2.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pyjwt = "^2.8.0"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
setuptools = "^69.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
